+++ relayer-cli/src/commands.rs
@@ -12,16 +12,14 @@ use abscissa_core::{Command, Configurable, FrameworkError, Help, Options, Runnab
 use crate::config::Config;
 
 use self::{
-    create::CreateCmds, keys::KeysCmd, light::LightCmd, listen::ListenCmd, query::QueryCmd,
-    start::StartCmd, start_multi::StartMultiCmd, tx::TxCmd, update::UpdateCmds,
-    version::VersionCmd,
+    create::CreateCmds, keys::KeysCmd, listen::ListenCmd, query::QueryCmd, start::StartCmd,
+    start_multi::StartMultiCmd, tx::TxCmd, update::UpdateCmds, version::VersionCmd,
 };
 use crate::commands::misbehaviour::MisbehaviourCmd;
 
 mod config;
 mod create;
 mod keys;
-mod light;
 mod listen;
 mod misbehaviour;
 mod query;
@@ -51,10 +49,6 @@ pub enum CliCmd {
     /// The `keys` subcommand
     #[options(help = "Manage keys in the relayer for each chain")]
     Keys(KeysCmd),
-
-    /// The `light` subcommand
-    #[options(help = "Basic functionality for managing the light clients")]
-    Light(LightCmd),
 
     /// The `create` subcommand
     #[options(help = "Create objects (client, connection, or channel) on chains")]