+++ relayer-cli/src/commands/listen.rs
@@ -2,7 +2,10 @@ use std::{ops::Deref, sync::Arc, thread};
 
 use abscissa_core::{application::fatal_error, error::BoxError, Command, Options, Runnable};
 use crossbeam_channel as channel;
+use itertools::Itertools;
 use tokio::runtime::Runtime as TokioRuntime;
+
+use tendermint_rpc::query::{EventType, Query};
 
 use ibc::ics24_host::identifier::ChainId;
 use ibc_relayer::{config::ChainConfig, event::monitor::*};
@@ -11,19 +14,30 @@ use crate::prelude::*;
 
 #[derive(Command, Debug, Options)]
 pub struct ListenCmd {
+    /// Identifier of the chain to listen for events from
     #[options(free)]
-    chain_id: Option<ChainId>,
+    chain_id: ChainId,
+
+    /// Add an event type to listen for, can be repeated. Listen for all events by default (available: Tx, NewBlock)
+    #[options(short = "e", long = "event", meta = "EVENT")]
+    events: Vec<EventType>,
 }
 
 impl ListenCmd {
     fn cmd(&self) -> Result<(), BoxError> {
-        let rt = Arc::new(TokioRuntime::new()?);
         let config = app_config();
 
-        let chain_id = self.chain_id.clone().unwrap();
-        let chain_config = config.find_chain(&chain_id).unwrap();
+        let chain_config = config
+            .find_chain(&self.chain_id)
+            .ok_or_else(|| format!("chain '{}' not found in configuration", self.chain_id))?;
 
-        listen(rt, chain_config.clone())
+        let events = if self.events.is_empty() {
+            &[EventType::Tx, EventType::NewBlock]
+        } else {
+            self.events.as_slice()
+        };
+
+        listen(chain_config, events)
     }
 }
 
@@ -35,11 +49,18 @@ impl Runnable for ListenCmd {
 }
 
 /// Listen to events
-pub fn listen(rt: Arc<TokioRuntime>, config: ChainConfig) -> Result<(), BoxError> {
-    info!(chain.id = %config.id, "spawning event monitor for");
+pub fn listen(config: &ChainConfig, events: &[EventType]) -> Result<(), BoxError> {
+    println!(
+        "[info] Listening for events `{}` on '{}'...",
+        events.iter().format(", "),
+        config.id
+    );
 
-    let (event_monitor, rx) = subscribe(config, rt)?;
-    let _ = thread::spawn(|| event_monitor.run());
+    let rt = Arc::new(TokioRuntime::new()?);
+    let queries = events.iter().cloned().map(Query::from).collect();
+    let (event_monitor, rx) = subscribe(&config, queries, rt)?;
+
+    thread::spawn(|| event_monitor.run());
 
     while let Ok(event_batch) = rx.recv() {
         dbg!(event_batch);
@@ -49,16 +70,22 @@ pub fn listen(rt: Arc<TokioRuntime>, config: ChainConfig) -> Result<(), BoxError
 }
 
 fn subscribe(
-    chain_config: ChainConfig,
+    chain_config: &ChainConfig,
+    queries: Vec<Query>,
     rt: Arc<TokioRuntime>,
 ) -> Result<(EventMonitor, channel::Receiver<EventBatch>), BoxError> {
-    let (mut event_monitor, rx) =
-        EventMonitor::new(chain_config.id, chain_config.websocket_addr, rt)
-            .map_err(|e| format!("couldn't initialize event monitor: {}", e))?;
+    let (mut event_monitor, rx) = EventMonitor::new(
+        chain_config.id.clone(),
+        chain_config.websocket_addr.clone(),
+        rt,
+    )
+    .map_err(|e| format!("could not initialize event monitor: {}", e))?;
+
+    event_monitor.set_queries(queries);
 
     event_monitor
         .subscribe()
-        .map_err(|e| format!("couldn't initialize subscriptions: {}", e))?;
+        .map_err(|e| format!("could not initialize subscriptions: {}", e))?;
 
     Ok((event_monitor, rx))
 }